Extract API into root folders
